fix: as service and guide
